move maximum sentry code to SDK from the client app to make sentry confiuration & setup minimal and easy in client apps
